+++ shared/CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 **Warning:** Features marked as *alpha* may change or be removed in a future release without notice. Use with caution.
 
-## [Unreleased]
+<!--## [Unreleased]-->
+
+## [2.2.0]
 
 ### Added
 
@@ -143,3 +145,4 @@ All notable changes to this project will be documented in this file.
 [2.0.0]: https://github.com/readium/r2-shared-swift/compare/2.0.0-beta.2...2.0.0
 [2.0.1]: https://github.com/readium/r2-shared-swift/compare/2.0.0...2.0.1
 [2.1.0]: https://github.com/readium/r2-shared-swift/compare/2.0.1...2.1.0
+[2.2.0]: https://github.com/readium/r2-shared-swift/compare/2.1.0...2.2.0